// backend/src/enums/feature-key.enum.ts
export enum FeatureKey {
   AUDIO_CALL = "audio_call",
   VIDEO_CALL = "video_call",
   SEND_MESSAGE = "send_message",
   PROFILE_BLUR = "profile_blur",
   MAX_INTERESTS = "max_interests",
   MAX_VIDEO_CALL_MINUTES = "max_video_call_minutes",
   MAX_AUDIO_CALL_MINUTES = "max_audio_call_minutes",
   MAX_MESSAGES = "max_messages",
}